requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
Django==6.0
fastapi==0.127.0
Flask==3.1.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
Werkzeug==3.1.4
celery
redis
python-dotenv
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
requests
django-filter
reportlab
django-redis
twilio
qrcode[pil]
google-auth
google-auth-httplib2
firebase_admin==7.4.0
